Update README.md
To get the most out of Stripchat as a viewer, here are some tips:

Create a self-ruling account. This allows you to yack with broadcasters, follow your favorites and get notifications when they come online.

Browse the categories and tags to find what interests you. There are so many options from amateurs to porn stars and everything in between. https://techytechtop.com/